refactor(capture): rejoin the extension-state and /sync clusters
Commit 58264122 ("split server internals into focused modules") minted extension_state_*
and sync_* in the same change, so the prefixes are provenance, not evidence. The reference
graph says the two prefixes are one strongly-connected cluster, not two:

  extension_state.go        -> sync.go            (mutual)
  extension_state_pilot.go  -> sync.go
  sync_state.go             -> extension_state.go
  sync_helpers.go           -> extension_state.go, extension_state_csp_security.go

and three of the six extension_state_* files have no outgoing edge at all — they were
carved out by field name, not by dependency. All twelve files mutate the same ~25
unexported ExtensionState fields under the same Capture.mu, which is why no subpackage
boundary is available here without exporting the whole struct (the internal/queries
lesson). Rejoining them along the real edges is the honest move.

  extension_state.go  <- _connection, _pilot, _tracking, _csp_security, _test_boundaries
  sync.go             <- sync_helpers.go, sync_timeout.go        (transport + policy)
  sync_state.go       <- sync_processing.go, sync_reconcile.go   (state mutation)

Transport and state mutation stay apart because that split *is* load-bearing: sync.go
runs outside the lock, sync_state.go runs inside it. Largest resulting file is 463 lines.

internal/capture: 53 -> 44 non-test files.
