+++ internal/capture/extension_state.go
@@ -1,10 +1,13 @@
-// Purpose: Defines extension connection/tracking/pilot/security state and related capture delegation methods.
-// Why: Centralizes extension lifecycle state used for routing, health checks, and command reconciliation.
+// Purpose: Owns ExtensionState — connection readiness, pilot gating, tab tracking, CSP posture and test boundaries.
+// Why: These fields share one lock and mutate together on every /sync round-trip; splitting them by prefix hid that.
 // Docs: docs/features/feature/backend-log-streaming/index.md
 
 package capture
 
-import "time"
+import (
+	"context"
+	"time"
+)
 
 const (
 	PilotStateAssumedEnabled     = "assumed_enabled"
@@ -91,4 +94,370 @@ type pilotStatusSnapshot struct {
 	Authoritative     bool
 	State             string
 	Source            string
+}
+
+// WaitForExtensionConnected polls until the extension connects, the timeout elapses,
+// or ctx is cancelled. Returns true if connected within the window, false otherwise.
+// A zero or negative timeout returns false immediately (no polling).
+// Poll interval is extensionReadinessPollInterval (200ms). ctx cancellation is
+// honoured between polls.
+func (c *Capture) WaitForExtensionConnected(ctx context.Context, timeout time.Duration) bool {
+	if c.IsExtensionConnected() {
+		return true
+	}
+	ctx, cancel := context.WithTimeout(ctx, timeout)
+	defer cancel()
+	ticker := time.NewTicker(extensionReadinessPollInterval)
+	defer ticker.Stop()
+	for {
+		select {
+		case <-ctx.Done():
+			return false
+		case <-ticker.C:
+			if c.IsExtensionConnected() {
+				return true
+			}
+		}
+	}
+}
+
+// IsExtensionConnected returns true if the extension has synced within the
+// disconnect threshold (10s). Returns false if never synced or stale.
+func (c *Capture) IsExtensionConnected() bool {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	return !c.extensionState.lastSyncSeen.IsZero() && time.Since(c.extensionState.lastSyncSeen) < extensionDisconnectThreshold
+}
+
+// GetExtensionStatus returns a detached connection snapshot.
+// Fields: connected (bool), last_seen (RFC3339 string), client_id (string).
+//
+// Failure semantics:
+// - If extension has never synced, last_seen is empty and connected=false.
+func (c *Capture) GetExtensionStatus() map[string]any {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+
+	connected := !c.extensionState.lastSyncSeen.IsZero() && time.Since(c.extensionState.lastSyncSeen) < extensionDisconnectThreshold
+
+	lastSeen := ""
+	if !c.extensionState.lastSyncSeen.IsZero() {
+		lastSeen = c.extensionState.lastSyncSeen.Format(time.RFC3339)
+	}
+
+	return map[string]any{
+		"connected": connected,
+		"last_seen": lastSeen,
+		"client_id": c.extensionState.lastSyncClientID,
+	}
+}
+
+// GetExtensionVersion returns the last reported extension version.
+func (c *Capture) GetExtensionVersion() string {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	return c.extensionState.extensionVersion
+}
+
+// GetVersionMismatch checks whether extension and server versions differ in major.minor.
+// Returns the extension version, server version, and whether a mismatch exists.
+// A mismatch is detected only when the extension has reported a version (non-empty)
+// and the major.minor portions differ from the server version.
+func (c *Capture) GetVersionMismatch() (extensionVersion string, serverVersion string, hasMismatch bool) {
+	c.mu.RLock()
+	extVer := c.extensionState.extensionVersion
+	srvVer := c.serverVersion
+	c.mu.RUnlock()
+
+	if extVer == "" || srvVer == "" {
+		return extVer, srvVer, false
+	}
+
+	extMajorMinor := majorMinor(extVer)
+	srvMajorMinor := majorMinor(srvVer)
+	if extMajorMinor == "" || srvMajorMinor == "" {
+		return extVer, srvVer, false
+	}
+
+	return extVer, srvVer, extMajorMinor != srvMajorMinor
+}
+
+// majorMinor extracts "X.Y" from a semver string "X.Y.Z".
+// Returns empty string if the version is not in a recognized format.
+func majorMinor(v string) string {
+	firstDot := -1
+	for i := 0; i < len(v); i++ {
+		if v[i] == '.' {
+			if firstDot == -1 {
+				firstDot = i
+			} else {
+				// Found second dot — return up to (but not including) it
+				return v[:i]
+			}
+		}
+	}
+	// No second dot found — not a valid semver with patch
+	if firstDot != -1 {
+		return v // "X.Y" format, return as-is
+	}
+	return ""
+}
+
+// IsPilotEnabled returns whether AI Web Pilot is currently enabled.
+func (c *Capture) IsPilotEnabled() bool {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	return c.extensionState.pilotEnabled
+}
+
+// IsPilotActionAllowed returns whether pilot-gated actions should be allowed.
+// Startup/reconnect uncertainty defaults to allowed until explicit disable arrives.
+func (c *Capture) IsPilotActionAllowed() bool {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	snap := pilotStatusSnapshotFromExtensionState(c.extensionState)
+	return snap.EffectiveEnabled
+}
+
+// GetPilotStatus returns pilot and heartbeat command status.
+// extension_connected uses the same threshold as IsExtensionConnected (10s on lastSyncSeen).
+// extension_last_seen is the RFC3339 timestamp of the last /sync, empty if never synced.
+//
+// Invariants:
+// - Returned in_progress slice is copied to prevent external mutation.
+func (c *Capture) GetPilotStatus() any {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	snap := pilotStatusSnapshotFromExtensionState(c.extensionState)
+
+	lastSeen := ""
+	if !c.extensionState.lastSyncSeen.IsZero() {
+		lastSeen = c.extensionState.lastSyncSeen.Format(time.RFC3339)
+	}
+
+	inProgress := make([]SyncInProgress, len(c.extensionState.inProgress))
+	copy(inProgress, c.extensionState.inProgress)
+	inProgressUpdated := ""
+	if !c.extensionState.inProgressUpdated.IsZero() {
+		inProgressUpdated = c.extensionState.inProgressUpdated.Format(time.RFC3339)
+	}
+
+	return map[string]any{
+		"enabled":             snap.EffectiveEnabled,
+		"configured_enabled":  snap.ConfiguredEnabled,
+		"authoritative":       snap.Authoritative,
+		"state":               snap.State,
+		"source":              snap.Source,
+		"extension_connected": !c.extensionState.lastSyncSeen.IsZero() && time.Since(c.extensionState.lastSyncSeen) < extensionDisconnectThreshold,
+		"extension_last_seen": lastSeen,
+		"in_progress_count":   len(inProgress),
+		"in_progress":         inProgress,
+		"in_progress_updated": inProgressUpdated,
+	}
+}
+
+// GetInProgressCommands returns a copy of latest extension-reported active commands.
+//
+// Failure semantics:
+// - Returns empty slice when no heartbeat data is available.
+func (c *Capture) GetInProgressCommands() []SyncInProgress {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	out := make([]SyncInProgress, len(c.extensionState.inProgress))
+	copy(out, c.extensionState.inProgress)
+	return out
+}
+
+// pilotStatusSnapshotFromExtensionState converts raw extension state to API-level pilot semantics.
+//
+// Failure semantics:
+// - Unknown/unset source fields fall back to conservative defaults.
+func pilotStatusSnapshotFromExtensionState(ext ExtensionState) pilotStatusSnapshot {
+	snap := pilotStatusSnapshot{
+		ConfiguredEnabled: ext.pilotEnabled,
+		Authoritative:     ext.pilotStatusKnown,
+	}
+
+	if !ext.pilotStatusKnown {
+		snap.EffectiveEnabled = true
+		snap.State = PilotStateAssumedEnabled
+		snap.Source = PilotSourceAssumedStartup
+		return snap
+	}
+
+	if ext.pilotEnabled {
+		snap.EffectiveEnabled = true
+		snap.State = PilotStateEnabled
+		if ext.pilotSource != "" {
+			snap.Source = ext.pilotSource
+		} else {
+			snap.Source = PilotSourceExtensionSync
+		}
+		return snap
+	}
+
+	snap.EffectiveEnabled = false
+	snap.State = PilotStateExplicitlyDisabled
+	if ext.pilotSource != "" {
+		snap.Source = ext.pilotSource
+	} else {
+		snap.Source = PilotSourceExtensionSync
+	}
+	return snap
+}
+
+// GetTrackingStatus returns the current tab tracking state.
+func (c *Capture) GetTrackingStatus() (enabled bool, tabID int, tabURL string) {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	return c.extensionState.trackingEnabled, c.extensionState.trackedTabID, c.extensionState.trackedTabURL
+}
+
+// UpdateTrackedTab programmatically updates the tracked tab state.
+// Used by switch_tab to retarget subsequent commands to the newly activated tab.
+//
+// Invariants:
+// - tabID must be > 0; zero/negative values are silently ignored.
+// - trackingEnabled is set to true when a valid tabID is provided.
+func (c *Capture) UpdateTrackedTab(tabID int, tabURL string, tabTitle string) {
+	if tabID <= 0 {
+		return
+	}
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	c.extensionState.trackingEnabled = true
+	c.extensionState.trackedTabID = tabID
+	c.extensionState.trackedTabURL = tabURL
+	c.extensionState.trackedTabTitle = tabTitle
+	c.extensionState.trackingUpdated = time.Now()
+}
+
+// GetTrackedTabTitle returns the tracked tab's title (may be stale).
+func (c *Capture) GetTrackedTabTitle() string {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	return c.extensionState.trackedTabTitle
+}
+
+// GetTabStatus returns the Chrome tab status ("loading", "complete", or empty if unknown).
+func (c *Capture) GetTabStatus() string {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	return c.extensionState.tabStatus
+}
+
+// IsTrackedTabActive returns whether the tracked tab is the foreground tab.
+// Returns (active, known). known=false means the extension has not reported this yet.
+func (c *Capture) IsTrackedTabActive() (bool, bool) {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	if c.extensionState.trackedTabActive == nil {
+		return false, false
+	}
+	return *c.extensionState.trackedTabActive, true
+}
+
+// SetTrackedTabActiveForTest sets the tracked tab active state for testing.
+func (c *Capture) SetTrackedTabActiveForTest(active bool) {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	c.extensionState.trackedTabActive = &active
+}
+
+// GetCSPStatus returns the last reported CSP restriction level for the tracked page.
+func (c *Capture) GetCSPStatus() (restricted bool, level string) {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	return c.extensionState.cspRestricted, c.extensionState.cspLevel
+}
+
+// CSPBlockedActions returns the actions blocked by the given CSP level and a
+// human-readable reason string. When the level is "none" or unrecognized, both
+// return values are nil/"" — callers should omit them from the response entirely
+// to avoid wasting tokens on normal pages.
+func CSPBlockedActions(level string) (actions []string, reason string) {
+	switch level {
+	case "script_exec":
+		return []string{"execute_js"},
+			"Page CSP blocks dynamic script execution. Use dom, get_readable, or list_interactive instead."
+	case "page_blocked":
+		return []string{
+				"execute_js", "click", "type", "select", "check", "scroll_to", "focus",
+				"get_text", "get_value", "get_attribute", "set_attribute",
+				"list_interactive", "get_readable", "get_markdown",
+				"fill_form", "fill_form_and_submit",
+			},
+			"Page blocks all script injection. Only navigate, screenshot, and network observation available."
+	default:
+		return nil, ""
+	}
+}
+
+// SetSecurityMode updates altered-environment mode reported to callers.
+// mode values: normal (default), insecure_proxy.
+//
+// Invariants:
+// - Any non-insecure mode value normalizes to SecurityModeNormal.
+// - Rewrite slice is copied on write to avoid external aliasing.
+func (c *Capture) SetSecurityMode(mode string, rewrites []string) {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+
+	switch mode {
+	case SecurityModeInsecureProxy:
+		c.extensionState.securityMode = SecurityModeInsecureProxy
+		c.extensionState.insecureRewrites = append([]string(nil), rewrites...)
+	default:
+		c.extensionState.securityMode = SecurityModeNormal
+		c.extensionState.insecureRewrites = nil
+	}
+}
+
+// GetSecurityMode returns current altered-environment mode and rewrite set.
+// production_parity is true only in normal mode.
+//
+// Invariants:
+// - Returned rewrite slice is copied and safe for caller mutation.
+func (c *Capture) GetSecurityMode() (mode string, productionParity bool, rewrites []string) {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+
+	mode = c.extensionState.securityMode
+	if mode == "" {
+		mode = SecurityModeNormal
+	}
+	productionParity = mode == SecurityModeNormal
+	rewrites = append([]string(nil), c.extensionState.insecureRewrites...)
+	return mode, productionParity, rewrites
+}
+
+// GetActiveTestIDs returns the list of currently active test IDs.
+func (c *Capture) GetActiveTestIDs() []string {
+	c.mu.RLock()
+	defer c.mu.RUnlock()
+	result := make([]string, 0, len(c.extensionState.activeTestIDs))
+	for testID := range c.extensionState.activeTestIDs {
+		result = append(result, testID)
+	}
+	return result
+}
+
+// SetTestBoundaryStart marks a test boundary as active for future event tagging.
+//
+// Invariants:
+// - activeTestIDs behaves as a set (idempotent insert).
+func (c *Capture) SetTestBoundaryStart(id string) {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	c.extensionState.activeTestIDs[id] = true
+}
+
+// SetTestBoundaryEnd clears a test boundary marker.
+//
+// Failure semantics:
+// - Deleting unknown IDs is a no-op.
+func (c *Capture) SetTestBoundaryEnd(id string) {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	delete(c.extensionState.activeTestIDs, id)
 }
