+++ internal/capture/sync.go
@@ -1,4 +1,4 @@
-// Purpose: Implements /sync transport flow for settings, logs, command results, and pending command delivery.
+// Purpose: Implements the /sync transport flow — settings, logs, command results, pending command delivery and long-poll policy.
 // Why: Consolidates extension-daemon synchronization into a single resilient protocol surface.
 // Docs: docs/features/feature/backend-log-streaming/index.md
 // Docs: docs/features/feature/query-service/index.md
@@ -8,9 +8,11 @@ package capture
 import (
 	"encoding/json"
 	"net/http"
+	"os"
 	"strings"
 	"time"
 
+	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/queries"
 	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/telemetry"
 	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/util"
 )
@@ -294,4 +296,65 @@ func (c *Capture) HandleSync(w http.ResponseWriter, r *http.Request) {
 	}
 
 	util.JSONResponse(w, http.StatusOK, resp)
+}
+
+// buildSyncCommands converts pending queries to sync commands.
+func buildSyncCommands(pending []queries.PendingQueryResponse) []SyncCommand {
+	commands := make([]SyncCommand, len(pending))
+	for i, q := range pending {
+		commands[i] = SyncCommand{
+			ID:            q.ID,
+			Type:          q.Type,
+			Params:        q.Params,
+			TabID:         q.TabID,
+			CorrelationID: q.CorrelationID,
+			TraceID:       q.TraceID,
+		}
+	}
+	return commands
+}
+
+// shouldEmitSyncSnapshot determines whether lifecycle telemetry should include a sync snapshot.
+func shouldEmitSyncSnapshot(req SyncRequest, state syncConnectionState, commandsOut int) bool {
+	if state.isReconnect || state.wasDisconnected || !state.wasConnected {
+		return true
+	}
+	if len(req.CommandResults) > 0 || commandsOut > 0 {
+		return true
+	}
+	if req.LastCommandAck != "" {
+		return true
+	}
+	return false
+}
+
+func (c *Capture) buildCaptureOverrides() map[string]string {
+	mode, productionParity, rewrites := c.GetSecurityMode()
+	if mode == SecurityModeNormal {
+		return map[string]string{}
+	}
+
+	overrides := map[string]string{
+		"security_mode":     mode,
+		"production_parity": "false",
+	}
+	if productionParity {
+		overrides["production_parity"] = "true"
+	}
+	if len(rewrites) > 0 {
+		overrides["insecure_rewrites_applied"] = strings.Join(rewrites, ",")
+	}
+	return overrides
+}
+
+const (
+	syncLongPollDefaultTimeout = 5 * time.Second
+	syncLongPollTestTimeout    = 100 * time.Millisecond
+)
+
+func syncLongPollTimeout() time.Duration {
+	if strings.HasSuffix(os.Args[0], ".test") {
+		return syncLongPollTestTimeout
+	}
+	return syncLongPollDefaultTimeout
 }
